APPEALS-48373 - Move tests out of DatePicker into QueueTable where they belong and add more tests
