+++ client/COPY.json
@@ -1574,5 +1574,6 @@
   "DATE_PICKER_DROPDOWN_AFTER": "After this date",
   "DATE_PICKER_DROPDOWN_ON": "On this date",
   "DATE_PICKER_CLEAR": "Clear filter",
-  "DATE_PICKER_DROPDOWN_LABEL": "Date filter parameters"
+  "DATE_PICKER_DROPDOWN_LABEL": "Date filter parameters",
+  "DATE_PICKER_QUICK_BUTTON_30": "Last 30 days"
 }
