+++ client/test/app/components/DatePicker.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen, fireEvent, waitFor, cleanup } from '@testing-library/react';
+import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 
 import {
   selectFromDropdown,
@@ -10,62 +10,6 @@ import {
 import { axe } from 'jest-axe';
 
 import DatePicker, { datePickerFilterValue } from 'app/components/DatePicker';
-import QueueTable from 'app/queue/QueueTable';
-import { encodeQueryParams } from 'app/util/QueryParamsUtil';
-import { when } from 'jest-when';
-import ApiUtil from 'app/util/ApiUtil';
-
-const records = [
-  { id: 1, date: '5/1/2024' },
-  { id: 2, date: '5/2/2024' },
-  { id: 3, date: '5/3/2024' },
-  { id: 4, date: '5/4/2024' },
-  { id: 5, date: '5/5/2024' },
-  { id: 6, date: '5/6/2024' },
-  { id: 7, date: '5/7/2024' },
-  { id: 8, date: '5/8/2024' },
-  { id: 9, date: '5/9/2024' },
-  { id: 10, date: '5/10/2024' },
-  { id: 11, date: '5/11/2024' },
-  { id: 12, date: '5/12/2024' },
-  { id: 13, date: '5/13/2024' },
-  { id: 14, date: '5/14/2024' },
-  { id: 15, date: '5/15/2024' },
-  { id: 16, date: '5/16/2024' },
-  { id: 17, date: '5/17/2024' },
-  { id: 18, date: '5/18/2024' },
-  { id: 19, date: '5/19/2024' },
-  { id: 20, date: '5/20/2024' }
-];
-
-const columns = [{
-  header: 'ID',
-  name: 'id',
-  enableFilter: false,
-  anyFiltersAreSet: false,
-  label: 'filter by id',
-  columnName: 'id',
-  valueName: 'ID',
-  valueFunction: (row) => row.id,
-  getSortValue: (row) => row.id
-}, {
-  header: 'Date',
-  name: 'date',
-  enableFilter: true,
-  anyFiltersAreSet: true,
-  label: 'filter by date',
-  columnName: 'date',
-  tableData: records,
-  valueName: 'Date',
-  valueFunction: (row) => row.date,
-  getSortValue: (row) => row.date,
-  filterType: 'date-picker',
-  filterSettings: {
-    buttons: false,
-    position: 'left'
-  },
-  customFilterMethod: datePickerFilterValue
-}];
 
 describe('DatePicker', () => {
   const handleChange = jest.fn();
@@ -90,17 +34,6 @@ describe('DatePicker', () => {
     await waitFor(() => {
       expect(screen.getByText('Date filter parameters')).toBeInTheDocument();
     });
-  };
-
-  const getTableDates = (container) => {
-    const tds = container.querySelectorAll('td');
-    let dates = [];
-
-    for (let i = 0; i < tds.length; i += 2) {
-      dates.push(tds[i + 1].innerHTML.trim());
-    }
-
-    return dates;
   };
 
   it('renders default state correctly', async () => {
@@ -176,271 +109,71 @@ describe('DatePicker', () => {
     expect(handleChange).toHaveBeenCalledWith('');
   });
 
-  describe('Front end filtering', () => {
+  it('has menu position left by default', async () => {
+    const { container } = setup();
 
-    const setupFrontend = () => {
+    openFilter(container);
 
-      return render(<QueueTable
-        columns={columns}
-        rowObjects={records}
-        enablePagination
-        casesPerPage={15}
-        anyFiltersAreSet
-        getKeyForRow={(_, row) => row.id}
-        defaultSort={{
-          sortColName: 'id',
-          sortAscending: true
-        }}
-      />);
-    };
-
-    it('renders in a queue table correctly', async () => {
-      const { container } = setupFrontend();
-
-      openFilter(container);
-
-      expect(container).toMatchSnapshot();
-
-    });
-
-    it('filters between dates', async () => {
-      const { container } = setupFrontend();
-
-      expect(screen.getAllByText('Viewing 1-15 of 20 total')[0]).toBeInTheDocument();
-
-      openFilter(container);
-
-      selectFromDropdown('Date filter parameters', 'Between these dates');
-
-      enterInputValue('start-date', '2024-05-01');
-      enterInputValue('end-date', '2024-05-10');
-
-      clickSubmissionButton('Apply Filter');
-
-      expect(screen.getAllByText('Viewing 1-10 of 10 total')[0]).toBeInTheDocument();
-
-      expect(getTableDates(container)).toEqual([
-        '5/1/2024',
-        '5/2/2024',
-        '5/3/2024',
-        '5/4/2024',
-        '5/5/2024',
-        '5/6/2024',
-        '5/7/2024',
-        '5/8/2024',
-        '5/9/2024',
-        '5/10/2024'
-      ]);
-    });
-
-    it('filters before date', async () => {
-      const { container } = setupFrontend();
-
-      expect(screen.getAllByText('Viewing 1-15 of 20 total')[0]).toBeInTheDocument();
-
-      openFilter(container);
-
-      selectFromDropdown('Date filter parameters', 'Before this date');
-
-      enterInputValue('start-date', '2024-05-10');
-
-      clickSubmissionButton('Apply Filter');
-
-      expect(screen.getAllByText('Viewing 1-9 of 9 total')[0]).toBeInTheDocument();
-
-      expect(getTableDates(container)).toEqual([
-        '5/1/2024',
-        '5/2/2024',
-        '5/3/2024',
-        '5/4/2024',
-        '5/5/2024',
-        '5/6/2024',
-        '5/7/2024',
-        '5/8/2024',
-        '5/9/2024'
-      ]);
-    });
-
-    it('filters after date', async () => {
-      const { container } = setupFrontend();
-
-      expect(screen.getAllByText('Viewing 1-15 of 20 total')[0]).toBeInTheDocument();
-
-      openFilter(container);
-
-      selectFromDropdown('Date filter parameters', 'After this date');
-
-      enterInputValue('start-date', '2024-05-15');
-
-      clickSubmissionButton('Apply Filter');
-
-      expect(screen.getAllByText('Viewing 1-5 of 5 total')[0]).toBeInTheDocument();
-
-      expect(getTableDates(container)).toEqual([
-        '5/16/2024',
-        '5/17/2024',
-        '5/18/2024',
-        '5/19/2024',
-        '5/20/2024'
-      ]);
-    });
-
-    it('filters on date', async () => {
-      const { container } = setupFrontend();
-
-      expect(screen.getAllByText('Viewing 1-15 of 20 total')[0]).toBeInTheDocument();
-
-      openFilter(container);
-
-      selectFromDropdown('Date filter parameters', 'On this date');
-
-      enterInputValue('start-date', '2024-05-15');
-
-      clickSubmissionButton('Apply Filter');
-
-      expect(screen.getAllByText('Viewing 1-1 of 1 total')[0]).toBeInTheDocument();
-
-      expect(getTableDates(container)).toEqual([
-        '5/15/2024'
-      ]);
-    });
-
-    it('allows the filter to be cleared', async () => {
-      const { container } = setupFrontend();
-
-      expect(screen.getAllByText('Viewing 1-15 of 20 total')[0]).toBeInTheDocument();
-
-      openFilter(container);
-
-      selectFromDropdown('Date filter parameters', 'On this date');
-
-      enterInputValue('start-date', '2024-05-15');
-
-      clickSubmissionButton('Apply Filter');
-
-      expect(screen.getAllByText('Viewing 1-1 of 1 total')[0]).toBeInTheDocument();
-
-      openFilter(container);
-
-      clickSubmissionButton('Clear filter');
-
-      expect(screen.getAllByText('Viewing 1-15 of 20 total')[0]).toBeInTheDocument();
-    });
-
+    expect(container.querySelectorAll('.date-picker.left').length).toBe(1);
   });
 
-  describe('Back end filtering', () => {
-    beforeEach(async () => {
-      ApiUtil.get = jest.fn();
+  it('allows setting the menu position', async () => {
+    const { container } = setup({ settings: { position: 'right' } });
 
-      when(ApiUtil.get).calledWith('/example/endpoint?&page=1').
-        mockResolvedValue(({
-          body: {
-            task_page_count: 1,
-            tasks: {
-              data: [
-                records[0],
-                records[1],
-                records[2],
-                records[3],
-                records[4]
-              ]
-            },
-            tasks_per_page: 15,
-            total_task_count: 5
-          }
-        }));
+    openFilter(container);
 
-      when(ApiUtil.get).calledWith('/example/endpoint?&page=1&filter%5B%5D=col%3Ddate%26val%3Don%2C2024-05-15%2C').
-        mockResolvedValue(({
-          body: {
-            task_page_count: 1,
-            tasks: {
-              data: [
-                records[0],
-                records[1],
-                records[2]
-              ]
-            },
-            tasks_per_page: 15,
-            total_task_count: 3
-          }
-        }));
+    expect(container.querySelectorAll('.date-picker.left').length).toBe(0);
+    expect(container.querySelectorAll('.date-picker.right').length).toBe(1);
+  });
 
+  it('has quick buttons off by default', async () => {
+    const { container } = setup();
+
+    openFilter(container);
+
+    expect(container.querySelectorAll('.quick-buttons').length).toBe(0);
+  });
+
+  it('allows enabling quick buttons', async () => {
+    const { container } = setup({ settings: { buttons: true } });
+
+    openFilter(container);
+
+    expect(container.querySelectorAll('.quick-buttons').length).toBe(1);
+  });
+
+  it('quick buttons can select last 30 days', async () => {
+    jest.spyOn(Date, 'now').mockReturnValue('2024-01-17T03:00:00.000-04:00');
+
+    const { container } = setup({ settings: { buttons: true } });
+
+    openFilter(container);
+
+    clickSubmissionButton('Last 30 days');
+
+    expect(handleChange).toHaveBeenCalledWith('between,2023-12-18,2024-01-17');
+  });
+
+  describe('datePickerFilterValue', () => {
+
+    it('returns true or false for between dates', async () => {
+      expect(datePickerFilterValue('5/15/2024', ['between,2024-05-01,2024-05-31'])).toBeTruthy();
+      expect(datePickerFilterValue('6/15/2024', ['between,2024-05-01,2024-05-31'])).toBeFalsy();
     });
 
-    afterEach(() => {
-      cleanup();
-      jest.clearAllMocks();
+    it('returns true or false for before date', async () => {
+      expect(datePickerFilterValue('4/15/2024', ['before,2024-05-01,'])).toBeTruthy();
+      expect(datePickerFilterValue('5/15/2024', ['before,2024-05-01,'])).toBeFalsy();
     });
 
-    const getUrlParams = (query) =>
-      Array.from(new URLSearchParams(query)).reduce((pValue, [kValue, vValue]) =>
-        Object.assign({}, pValue, {
-          [kValue]: pValue[kValue] ? (Array.isArray(pValue[kValue]) ?
-            pValue[kValue] : [pValue[kValue]]).concat(vValue) : vValue
-        }), {});
-
-    const setupBackend = () => {
-      return render(<QueueTable
-        columns={columns}
-        rowObjects={[]}
-        enablePagination
-        casesPerPage={15}
-        useTaskPagesApi
-        taskPagesApiEndpoint={`/example/endpoint${encodeQueryParams()}`}
-        anyFiltersAreSet
-        tabPaginationOptions={getUrlParams(window.location.search)}
-        getKeyForRow={(_, row) => row.id}
-        skipCache
-      />
-      );
-    };
-
-    it('renders in a queue table correctly', async () => {
-      const { container } = setupBackend();
-
-      await waitFor(() =>
-        expect(screen.getAllByText('Viewing 1-5 of 5 total')[0]).toBeInTheDocument()
-      );
-
-      openFilter(container);
-
-      expect(container).toMatchSnapshot();
-
+    it('returns true or false for after date', async () => {
+      expect(datePickerFilterValue('5/15/2024', ['after,2024-05-01,'])).toBeTruthy();
+      expect(datePickerFilterValue('4/15/2024', ['after,2024-05-01,'])).toBeFalsy();
     });
 
-    it('updates url params correctly on apply and reset', async () => {
-      const { container } = setupBackend();
-
-      expect(ApiUtil.get).toHaveBeenCalledWith('/example/endpoint?&page=1');
-
-      await waitFor(() =>
-        expect(screen.getAllByText('Viewing 1-5 of 5 total')[0]).toBeInTheDocument()
-      );
-
-      openFilter(container);
-
-      selectFromDropdown('Date filter parameters', 'On this date');
-
-      enterInputValue('start-date', '2024-05-15');
-
-      clickSubmissionButton('Apply Filter');
-
-      expect(ApiUtil.get).toHaveBeenCalledWith(
-        '/example/endpoint?&page=1&filter%5B%5D=col%3Ddate%26val%3Don%2C2024-05-15%2C');
-
-      await waitFor(() =>
-        expect(screen.getAllByText('Viewing 1-3 of 3 total')[0]).toBeInTheDocument()
-      );
-
-      openFilter(container);
-
-      clickSubmissionButton('Clear filter');
-
-      expect(ApiUtil.get).toHaveBeenLastCalledWith('/example/endpoint?&page=1');
-
+    it('returns true or false for on date', async () => {
+      expect(datePickerFilterValue('5/1/2024', ['on,2024-05-01,'])).toBeTruthy();
+      expect(datePickerFilterValue('5/2/2024', ['on,2024-05-01,'])).toBeFalsy();
     });
-
   });
 });
